Batterieboost spec new
